port wip

+++ endpoints_management/auth/tokens.py
@@ -16,17 +16,19 @@
 
 from __future__ import absolute_import
 
+from past.builtins import basestring
+from builtins import object
 import datetime
-import jwkest
+#import jwkest
 import time
 
 from dogpile import cache
-from jwkest import jws
-from jwkest import jwt
+#from jwkest import jws
+#from jwkest import jwt
 
 from . import suppliers
 
-INT_TYPES = (int, long)
+INT_TYPES = (int, int)
 
 
 class Authenticator(object):  # pylint: disable=too-few-public-methods
